+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 3.5.2
 ----------------
 - Fix pg_connect() error in database.inc.php
+- Fix DB error with REPORTING_GP_SCALE field numeric(10,3) type in Schools.php
 
 Changes in 3.5.1
 ----------------